requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-heroku==0.3.1
geoip2==3.0.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
ipinfo==3.0.0
lxml==4.6.1
maxminddb==1.5.4
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
public==2019.4.13
python3-openid==3.1.0
pytz==2020.1
query-string==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.0.1
